feat: remove files and move it on Client file
